requirements.txt
streamlit>=1.32
PyPDF2>=3.0.0
scikit-learn>=1.3.0
pdfminer.six>=20221105
